requirements.txt
aiohttp==3.8.4
deepl==1.14.0
gidgethub==5.2.1
httplib2==0.22.0
matplotlib==3.7.1
oauth2client==4.1.3
parsedatetime==2.6
pydrive==1.3.1
unidecode==1.3.6
tzdata==2023.3
APScheduler==3.10.1
revolt.py==0.1.11